put timer component outside of quiz component to prevent rerendering and stopping time
