+++ src/Components/Quiz.js
@@ -1,13 +1,10 @@
 import React, { useState, useEffect } from "react";
-import { Modal, Button } from "react-bootstrap";
 import Question from "./Questions";
-import Timer from "./Timer";
 import firebase from "../Firebase";
 const Quiz = (props) => {
   const db = firebase.firestore();
 
   const quizRef = db.collection("quiz");
-  const statRef = db.collection("stats");
 
   const [codequiz, setCodeQuiz] = useState([]);
   const [quizcounter, setquizCounter] = useState({
@@ -20,11 +17,6 @@ const Quiz = (props) => {
     streak: 0,
     nickname: "",
   });
-
-  // boostrap modal states
-  const [show, setShow] = useState(false);
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
 
   useEffect(() => {
     quizRef
@@ -41,6 +33,16 @@ const Quiz = (props) => {
       });
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  useEffect(() => {
+    if (props.timerstatus === false) {
+      props.setquizpoints({
+        ...props.quizentry,
+        points: points.quizpoints,
+      });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [props.timerstatus]);
 
   // checks if answer is correct | will add points to if correct
   const checkData = (answer) => {
@@ -80,24 +82,6 @@ const Quiz = (props) => {
     }
 
     hideData();
-  };
-
-  // saves highscore with nickname and sends back to homescreen
-  const save = () => {
-    const date = new Date();
-    const day = date.getDate();
-    const month = date.getMonth();
-    const year = date.getFullYear();
-
-    // const hours = date.getHours();
-    // const minutes = date.getMinutes();
-
-    statRef.doc().set({
-      nickname: points.nickname,
-      score: points.quizpoints,
-      date: `${month}-${day}-${year}`,
-    });
-    props.setquiz(false);
   };
 
   // setTimeout on displaying status of answer
@@ -143,7 +127,7 @@ const Quiz = (props) => {
           View high scores
         </button>
         <h2 style={{ textShadow: "2px 2px 2px gray" }}>codequiz</h2>
-        <Timer handleshow={handleShow} quizcounter={quizcounter} />
+        {/* <Timer handleshow={handleShow} /> */}
       </div>
       <hr />
       <div style={{ textAlign: "center" }}>
@@ -178,41 +162,6 @@ const Quiz = (props) => {
         <br />
         <h3 style={displayAnswer}>{quizcounter.answer}</h3>
       </div>
-      <>
-        <Modal show={show} onHide={handleClose}>
-          <Modal.Header closeButton>
-            <Modal.Title>Final Results</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            Wooo, you got {points.quizpoints} points!{" "}
-            <form>
-              <input
-                style={{ padding: "5px", borderRadius: "10px" }}
-                placeholder="Enter your nickname"
-                onChange={(e) =>
-                  setPoints({ ...points, nickname: e.target.value })
-                }
-              />
-            </form>
-          </Modal.Body>
-          <Modal.Footer>
-            <Button
-              style={{
-                borderRadius: "5px",
-                padding: "10px",
-                color: "white",
-                backgroundColor: "purple",
-                borderColor: "purple",
-                boxShadow: "2px 3px 4px gray",
-                cursor: "pointer",
-              }}
-              onClick={save}
-            >
-              Submit
-            </Button>
-          </Modal.Footer>
-        </Modal>
-      </>
     </div>
   );
 };
